Revert "Merge branch 'income-expense-tracker-api' into 'master'"
This reverts merge request !1

+++ src/main/java/com/finadv/controller/UserController.java
@@ -45,7 +45,7 @@ public class UserController {
 	 * @return
 	 */
 	@GetMapping("/users/{id}")
-	public ResponseEntity<?> showUser(@PathVariable Long id) {
+	public ResponseEntity<?> showUser(@PathVariable int id) {
 
 		User user = userService.getUserById(id);
 		if (user != null)
@@ -69,7 +69,7 @@ public class UserController {
 	 * @return
 	 */
 	@PutMapping("/users/{id}")
-	public ResponseEntity<?> updateUser(@RequestBody User user, @PathVariable Long id) {
+	public ResponseEntity<?> updateUser(@RequestBody User user, @PathVariable int id) {
 		user.setId(id);
 		User updatedUser = userService.updateUser(user);
